debugged

+++ Templates/uh_script.sh
@@ -1,6 +1,6 @@
 #!/bin/sh -e
 
-sleep 30
+date >  /beenhere.txt
 
 if [ -f /boot/log2ram.mk ]
 then
@@ -9,15 +9,16 @@ then
   echo "log2ram installed" >> /uh.log
 fi
 
-echo "" >> /uh.log
+echo "" > /uh.log
 echo "" >> /uh.log
 echo "" >> /uh.log
 
 if [ ! -f /boot/chown_pi ]
+then
   echo "Changing ownership of /home/pi ." >> /uh.log
   chown -R pi:pi /home/pi >> /uh.log
-  touch /boot/chown_pi
-then
+  touch /boot/chown_pi_done
+
 
 echo "" >> /uh.log
 echo "" >> /uh.log
@@ -36,15 +37,16 @@ then
   # the presence of /boot/resized marks that the resize effort took already place
   if [ ! -f /boot/resized ]
   then
-    echo "preparing for resizing file system while booting next time." >> /uh.log
-    cat /boot/cmdline_resize.txt > /boot/cmdline.txt 2>&1 >> /uh.log
+    if [ -f /boot/cmdline_resize.txt ]; then
+      echo "preparing for resizing file system while booting next time." >> /uh.log
+      cat /boot/cmdline_resize.txt > /boot/cmdline.txt 2>&1 >> /uh.log
   
-    touch /boot/resized
-
-    echo "About to reboot ..." >> /uh.log
-    /sbin/reboot
+      touch /boot/resized
+    else
+      echo "The file /boot/cmdline_resize.txt is missing. Cannot prepare for resize. The file /boot/noresize will be created again." >> /uh.log
+      touch /boot/noresize
+    fi
   fi
 fi
 
-exit 0
